Allow `L<text|schem:...>` tags.
As of the latest `perlpod` and `perlpodspec` committed to blead and forthcoming in Perl 5.12, and now included in Pod::Simple 3.11, on CPAN, the  formerly disallowed `L<text|schem:...>` tags are now allowed. Pod::Simple 3.11 has been updated to take full advantage of this format in all output formats (including plain text). Thus, it makes sense for Pod::Test to now allow this tag. It also simplifies the code a bit. :-)