Fixed a test in the internal _is_perl method to test for a (suspected) Perl file.
The POD documentation stated that a Perl file is, among the others, "any file that ends in .bat and has a first line with --*-Perl-*--", however the _is_perl method was testing only for the latter condition, skipping the .bat part.
Now the _is_perl method tests that the file has either a shee-bang or the right well-known Perl suffix OR ends with .bat and has the Perl comment line as first.
The test for the well known Perl extension has been pushed to the end of the method, so that if a file cannot be read (i.e., open fails) the file is not considered to be a good Perl file (e.g., it is not readable).