Improved the _is_perl internal method following the suggestions of David, see https://github.com/theory/test-pod/pull/7
As suggested, the method should accept any file with a good suffix before trying to open it, we are sure it is a file from the finding subroutine.
All regular expressions are now using non-capturing groups, since there is no back reference used at all within any scope. All regular expressions dealing with file name (extensions) use the /x capability for better readibility.
If a file has not a well know suffix it will be tested against a shee
bang in the first line or, if it has a .bat suffix, it will be tested
against a --*-Perl-*-- first line.
Note that the MS Windows batch file extension is checked with the
ignore-case flag, while the *nix-like well know suffixes are not, since
it make sense to distinguish good cases .pl and .PL but not .Pl.